feat: move receipt impl to v0.6 file
Since it implements the v0.6 structure by default.
