+++ crates/rpc/src/v06.rs
@@ -1,6 +1,6 @@
 use crate::jsonrpc::{RpcRouter, RpcRouterBuilder};
 
-mod method;
+pub(crate) mod method;
 pub(crate) mod types;
 
 use crate::v02::method as v02_method;
